New exception detected and implemented: Anglicism
Doing random testing I found the case of words that comes from
English and ends in "y". Now it's implemented.

Code may need refactoring. It's not out of control for me now, but
maybe it will be in few months. There are methods (like process) that
are "god methods" and does too many things, it may be breaking SRP.

I placed some additional comments to help.